# openapi/openapi.yml
openapi: 3.0.1
info:
  title: Finance App
  description: Finnance application
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: Kaloyan Velichkov
    email: kaloyanvellichkov@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.3
servers:
  - url: 'https://financeapp-be.vercel.app/'
tags:
  - name: Income
    description: Provides a set of functions to interact with the income data
  - name: Expense
    description: Provides a set of functions to interact with the expense data
  - name: User
    description: Provides a set of functions to manage the user data and session
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /income/get:
    get:
      tags:
        - Income
      summary: Вземане на приходи
      description: Връща приходите за даден потребител
      operationId: GetIncome
      parameters:
        - in: query
          name: fromDate
          schema:
            type: string
            format: date-time
            example: '2022-10-01T13:50:00.000Z'
        - in: query
          name: toDate
          schema:
            type: string
            format: date-time
            example: '2022-10-31T13:50:00.000Z'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - userId
                    - date
                    - amount
                    - comment
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    date:
                      type: string
                      format: date-time
                      example: '2022-10-09T13:50:00.000Z'
                    amount:
                      type: number
                      example: 2300.2
                    comment:
                      type: string
                      example: Вечеря
        '500':
          description: Error response
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    description: description of the error
  /income/add:
    post:
      tags:
        - Income
      summary: Добавяне на приход
      description: Добавя приход
      operationId: AddIncome
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - amount
                - comment
              properties:
                date:
                  type: string
                  format: date-time
                  example: '2022-10-09T13:50:00.000Z'
                amount:
                  type: number
                  example: 2300.2
                comment:
                  type: string
                  example: Заплата
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  '/income/update/{id}':
    put:
      tags:
        - Income
      summary: Редакторане на приход
      description: Редактира приход
      operationId: UpdateIncome
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Income id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                  example: '2022-10-09T13:50:00.000Z'
                amount:
                  type: number
                  example: 600.8
                comment:
                  type: string
                  example: Продадох си телефона
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  '/income/delete/{id}':
    delete:
      tags:
        - Income
      summary: Изтриване на приход
      description: Изтрива приход
      operationId: DeleteIncome
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Income id
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  /expense/get:
    get:
      tags:
        - Expense
      summary: Вземане на разходи
      description: Връща разходите за даден потребител
      operationId: GetExpense
      parameters:
        - in: query
          name: fromDate
          schema:
            type: string
            format: date-time
            example: '2022-10-01T13:50:00.000Z'
        - in: query
          name: toDate
          schema:
            type: string
            format: date-time
            example: '2022-10-31T13:50:00.000Z'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - userId
                    - date
                    - amount
                    - comment
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    date:
                      type: string
                      format: date-time
                      example: '2022-10-09T13:50:00.000Z'
                    amount:
                      type: number
                      example: 2300.2
                    comment:
                      type: string
                      example: Вечеря
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  /expense/add:
    post:
      tags:
        - Expense
      summary: Добавяне на разход
      description: Добавя разход
      operationId: AddExpense
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - amount
                - comment
              properties:
                date:
                  type: string
                  format: date-time
                  example: '2022-10-09T13:50:00.000Z'
                amount:
                  type: number
                  example: 2300.2
                comment:
                  type: string
                  example: Вечеря
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  '/expense/update/{id}':
    put:
      tags:
        - Expense
      summary: Редакторане на разход
      description: Редактира разход
      operationId: UpdateExpense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Expense id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                  example: '2022-10-09T13:50:00.000Z'
                amount:
                  type: number
                  example: 150.8
                comment:
                  type: string
                  example: Гориво
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  '/expense/delete/{id}':
    delete:
      tags:
        - Expense
      summary: Изтриване на разход
      description: Изтрива разход
      operationId: DeleteExpense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Expense id
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  /user/register:
    post:
      tags:
        - User
      summary: Регистриране на потребител
      description: Регистрира потребител
      operationId: RegisterUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - password
              properties:
                email:
                  type: string
                  example: example@mail.com
                name:
                  type: string
                  example: Pesho Shvepsa
                password:
                  type: string
                  example: my_password_123
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  /user/login:
    post:
      tags:
        - User
      summary: Логване на потребител
      description: Логва потребител
      operationId: LoginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: example@mail.com
                password:
                  type: string
                  example: my_password_123
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - email
                  - name
                  - jwtToken
                properties:
                  id:
                    type: string
                  email:
                    type: string
                    example: example@mail.com
                  name:
                    type: string
                    example: Pesho Shvepsa
                  jwtToken:
                    type: string
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'
  '/user/update/{id}':
    put:
      tags:
        - User
      summary: Редакторане на потребител
      description: Редактира потребител
      operationId: UpdateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: example@mail.com
                name:
                  type: string
                  example: Pesho Shvepsa
                password:
                  type: string
                  example: my_password_123
      responses:
        '200':
          description: OK
        '500':
          $ref: '#/paths/~1income~1get/get/responses/500'